+++ src/model/Towers/Tower.h
@@ -93,7 +93,7 @@ protected:
     bool isCurrentTargetOutOfRange(const std::vector<Enemy*>&) const;
     void changeTarget(const std::vector<Enemy*>&);
     void hitCurrentTarget(unsigned int dmg);
-    int getRangeInTileSizes(int range);
+    int getRangeInTileSizes(float range);
 };
 
 #endif //TOWERDEFENSE_TOWER_H
